backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Environmental monitoring assistant powered by Ollama Phi-3"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
    "psycopg2-binary>=2.9.11",
    "pandas>=3.0.1",
    "joblib>=1.5.3",
    "scikit-learn>=1.8.0",
    "ollama>=0.6.1",
    "uvicorn>=0.41.0",
    "python-multipart>=0.0.22",
    "supabase>=2.28.2",
]
